+++ jsonargparse_tests/cli_tests.py
@@ -24,6 +24,34 @@ class CLITests(unittest.TestCase):
         self.assertEqual('Y', CLI(functions, args=['cmd2', '--a2=Y']))
 
 
+    def test_single_class_cli(self):
+        class Class1:
+            def __init__(self, i1: str):
+                self.i1 = i1
+            def method1(self, m1: bool):
+                return self.i1, m1
+
+        self.assertEqual(('1', False), CLI(Class1, args=['1', 'method1', 'false']))
+
+
+    def test_function_and_class_cli(self):
+        def cmd1(a1: int):
+            return a1
+
+        class Cmd2:
+            def __init__(self, i1: str = 'd'):
+                self.i1 = i1
+            def method1(self, m1: float):
+                return self.i1, m1
+            def method2(self, m1: int = 0):
+                return self.i1, m1
+
+        components = [cmd1, Cmd2]
+        self.assertEqual(5, CLI(components, args=['cmd1', '5']))
+        self.assertEqual(('d', 1.2), CLI(components, args=['Cmd2', 'method1', '1.2']))
+        self.assertEqual(('b', 3), CLI(components, args=['Cmd2', '--i1=b', 'method2', '--m1=3']))
+
+
     def test_empty_context(self):
         def empty_context():
             CLI()
@@ -32,13 +60,23 @@ class CLITests(unittest.TestCase):
 
 
     def test_non_empty_context(self):
-        def non_empty_context():
+        def non_empty_context_1():
             def function(a1: float):
                 return a1
 
             return CLI(args=['6.7'])
 
-        self.assertEqual(6.7, non_empty_context())
+        def non_empty_context_2():
+            class Class1:
+                def __init__(self, i1: str):
+                    self.i1 = i1
+                def method1(self, m1: int):
+                    return self.i1, m1
+
+            return CLI(args=['a', 'method1', '2'])
+
+        self.assertEqual(6.7, non_empty_context_1())
+        self.assertEqual(('a', 2), non_empty_context_2())
 
 
 if __name__ == '__main__':